Enabled 128-bit masks.
I added the option to compile using 128-bit masks instead of the
native 64-bit masks.  The compiler simulates the 128-bit type
with two 64-bit words, so this is slower than just using the
64-bit masks.  However, it enables handling larger graphs.

I used a macro to determine which type to use for the bit masks.
This macro can be used from the Makefile.  "make all" makes the
usual precolor_extend executable, as well as precolor_extend128.