Changing how we select the maximum color when advincing to the next vertex.
For extension vertices, always use max_num_colors.  We do not try to
restrict to just one new color used, nor deal with vertices in orbit
with the previous vertex.  This simplifies the code, as well as makes it
easier to re-use previous extension colorings.

For precolored vertices, we use at most one new color, and we handle
vertices in orbit with the previous vertex.  We also eliminate the use
of mask_skip_max_color_to_try.